refactor: add Navigation class
